requirements.txt
# Web 框架
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# LangChain 生态
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.0.1

# 向量数据库
chromadb>=0.4.0

# 知识图谱
neo4j>=5.0.0

# 搜索
duckduckgo-search>=3.0.0

# AI API
zhipuai>=2.0.0
requests>=2.28.0

# 工具
python-dotenv>=1.0.0
